.npmrc: Move to monorepo root
I wanted to only have this affect the backend, but it seems that
putting it in the `backend` directory causes it to be ignored when the
installation is run from the monorepo root. Putting it here causes it
to take effect in the easy case, where we run `pnpm install` from the
root.