Strapi: cache images for a year
Strapi's support for "replacing" media involves uploading it to the
exact same url. If we allow this, we have to turn caching down to hours
instead of days or months, significantly reducing the utility of
caching.

So, let's discourage that method by making it not work. Cache images for
a year. If you want to change an image, upload a new one, don't use the
built-in "replace media" button.