Parts Pages: load images for image-less strapi nodes
The list of child-nodes we show on a given product list page (parent
device) sometimes don't match up with the list of device children in our
ifixit.com hierarchy. When we query ifixit.com for the parent device, we
only get pictures for the immediate children. For those nodes in strapi
which aren't immediate children of the parent-device, the ifixit.com api
thus won't give us pictures. For these nodes without pictures, query
their information (in bulk) via the new topic_images API so we can get
images for them.