fix: Split prod/dev Sentry reporting in /frontend
We want to report dev events to react-commerce-dev in Sentry, and prod
events to react-commerce in Sentry.

By default, we'll set the sample rate to 0 for dev so that CI failures
don't report to Sentry.

This also adds a `next.runtime` tag to /frontend events that helps
quickly differentiate between events that happen on the server versus
the client. There is a pre-existing `runtime` tag that uses 'browser'
for the client and 'node' for the server, but we'll be a little more
specific with our tag.