feat: Use Strapi Sentry plugin in /backend
https://strapi.io/blog/introducing-sentry-plugin

According to this blog, the primary purpose of this plugin is NOT to report
errors when the Strapi server fails to start, but to instead send errors when
there are problems handling requests.

> Send errors encountered in your application's end API to Sentry

I only saw one `.env` file, so I didn't include the SENTRY_DSN environment var
here (which differs between prod and dev). That is handled by JS instead.