sentry: Get DSN from environment variable
Instead of hard coding a fallback. This is simpler, and we can always
set the sample rate to 1 because if SENTRY_DSN is undefined then we know
errors won't get reported.

I put the react-commerce-dev DSN in some of the example .env files
because it's ok for these to be public according to Sentry:
https://docs.sentry.io/product/sentry-basics/dsn-explainer/#dsn-utilization