iFixitImage: Use `unoptimized` prop
Turns out, if you use `loader=..` next/image prints a warning if your
function doesn't make use of the `width` prop. It tells you to use
unoptimized=true if you just want to not use the image pipeline.

So, let's do that for now until we decide which images to resample /
recompress.