useAuthenticatedUser(): never refetch
If we've retrieved the user data, don't refetch even if the component is
unmounted / remounted.

useQuery() will, by default, refetch on mount / window focus / ...
several other events *if* the data is considered stale. Another default
treats *all* data as stale by default.

This was causing a problem because re-rendering the InstantSearch
component causes it to unmount. Unmounting causes a react-query refetch
on remounting, which causes a re-render and so-on.