Algolia API Key: use user-specific API key if available
If a user ia pro, they are allowed to see "pro" products and thus need a
different API key than the default one that hides pro products. Our API
will include a different API key if one applies to the current user.

This seems simple enough, but it seems <InstantSearch> effectively has a
bug. If you re-render it with a different search client (like say with a
different API key) it doesn't re-run the search query and continues to
show the same SSR results. It only runs the seach query if your
search query itself changes (new filters, different query, ...).

It seems there's no easy way to tell it to re-run the current search,
nor to *not* render the SSR results withotu manually setting the
server-side state to null. Unfortunately, doing so temporarily renders
"blank"