Algolia Api Key: rerun search when changing key
Algolia InstantSearch doesn't consider the API key to part of the search
query so the cached search results aren't considered invalid when
only changing the API key.

Let's "fix" this bug by refreshing the search results if the api key
changes. This is a bit tricky cause we *really* don't want it to happen
multiple times. It also needs to be *under* the <InstantSearch>
component, but that is unmounted / remounted so we can't use any of refs
/ effects / memo / state to only run once.

Solution is to make a hook that generates a component which updates the
ref after refreshing.

It now only runs the query once, or not at all, so success!

Still need to experiment to see if this dynamic component solution
re-renders more often or something.