iFixit Origin: use relative urls if on same origin
Effectively, if we are running the app on the same base domain as the
urls of the API, then we should use relative urls in for all front-end
API requests.

Backend requests still need to the full url because there is no "origin"
as they aren't in the context of a browser.

Determine if we should use relatively urls by comapring the base domain
of IFIXIT_ORIGIN and the base domain of the current url. We are using
base domain here so we allow using relative urls when on language
subdomains like es.ifixit.com (where IFIXIT_ORIGIN === www.ifixit.com).