Add tests for /Parts page devices.
This is a similar replica of WDStorePagesSmokeTest in iFixit/ifixit.
The first test visits Parts page and checks if the first few device
pages display fine. We can update this test when the seed data is updated.

The second test clicks on a single device until no nested devices are there.
Then makes sure that there are at least one product.