fix: Hide refinement items whose count equals hits
If a refinement item (aka a facet option) is applicable to the same
number of products as the number of products in the current view, then
that refinement item is essentially a no-op and we should hide it.

One exception: Refinement items that are currently selected should still
be displayed.

Extracts this logic into a hook so we can use it to decide whether an entire
refinement list should be rendered, and whether each refinement list item
should be rendered.

I noticed that the `hits` from `useHits()` seemed to only include the hits for
the current page, because `hits.length` was consistently 20. I opted to use
`results.nbHits` instead, which reflects the total number of hits for the
current search (which can be far greater than 20).