facets: Scroll mobile drawer to top
Scrolls the mobile facets drawer to the top whenever a facet is selected, or the
"Go Back" button is pressed.

Fixes a UX bug where the user could still be at the bottom of the page after
navigating between panels of the drawer. This sometimes caused a blank screen if
there weren't enough facets/items to reach the scroll position of the viewport.