instantsearch: Add custom url handling
Configures the InstantSearchSSRProvider to use our code to create/parse
urls given the state of the current search. Follows the example
implementation given in Algolia's docs:
https://www.algolia.com/doc/guides/building-search-ui/going-further/routing-urls/react-hooks/#seo-friendly-urls

This is all so we can create links to filtered collection pages in our
code. Namely, in `FacetFilter` we'd like to link to Item Type pages for
SEO reasons, but we also want to carry the current filters forward when
the user navigates to those pages. To build those links, we use the
state of the current search.
https://www.algolia.com/doc/api-reference/widgets/use-instantsearch/react-hooks/