fix: Resolve InstantSearch type errors
We need to tell typescript that InstantSearch uses our `RouteState`
custom type. Before, it was typed as `any` and we were getting errors
when we tried to access certain keys. It feels a little disjointed
assiging `routing` to a variable, then using it in the component, but I
didn't know how else to type it.

Also drops the `routing` prop to the provider, we weren't using it
anyways.