fix: `Item Type` can be a query param
On the /Parts page, we'd still like to be able to filter by Item Type,
but it should end up as a query param, not as the slug.

This updates createURL() to only omit `Item Type` from the filters if
there is a deviceHandle in the url. Otherwise, treat `Item Type` like
any other filter. parseURL() already does this check.

In FacetFilter, ignore `Item Type` because we only make links for
deviceHandle pages here.