item type: Make it a single select facet
Introduces the notion of single and multi select to facets. Multi select
facets will render a refinement list as before (with checkboxes). Single
select facets will render the RefinementMenu. The entire facet panel for
single select facets will disappear when they are selected, and
disappear when that filter is deselected.

There should be no change in behavior besides the
disappearing/reappearing of the `Item Type` facet when an item type is
selected.