Modify the css properties to get rid of the empty tile.
Previously, we were setting the grid background to gray and that
was causing the weird empty tile if a product wasn't displayed.
- This commit gets rid of the bg color for the grid and then adds
borders for each individual link box.
-This ensures that if there is no product, there is no link box
and hence no border.
   - The reason border was added to the link box was because previosly,
the bg color for the grid formed the border to differentiate b/w the
link boxes. Now that the bg color is removed, the border properties
were added to keep the style for the rest of the page consistent.