Derive Item Type from InstantSearch state
Before, the item type was constant for every page. Now, it can change
based on whether the user selects/deselects the item type facet, and the
page won't reload.

Introduces a new hook to get the currently selected item type for a
device parts page. This hook is used in places where
`productList.itemType` was used before.