fix: Don't filter `Item Type` facet
We already handle the `Item Type` facet in FacetsAccordian and
FacetsDrawer, so we don't need to filter out `Item Type` in the
useFilteredFacets hook.

This fixes "unmounted component" errors that were showing up in the
console. It also fixes a bug where, when an item type was selected on a
device parts page, the page would flash to the item type page, then back
to the device parts page without the refinement applied.