fix: Drop `<NextLink>` on product list children
There was a bug where, after clicking on a product list child link, the
new page would render with empty results invariably. When you reloaded
the page, the results would appear.

I'm not sure which commit in this pull introduced the bug. However, it
vanishes when you drop NextLink here, so my guess is that Algolia's
routing is not working with Next.js's routing here, as mentioned in
`<InstantSearchProvider>`.