url: Properly encode item type handle
The device handle always uses underscores in the URL. This change makes
the item type handle follow the same pattern, where underscores are
always used in the url and spaces are used in code.

Both "/Parts/Device/Headphone Jack" and "/Parts/Device/Headphone_Jack"
will serve the correct page, and the space in the former will be
converted to an underscore.