fix: Always remount InstantSearchProvider
Reverts 45239330b8249bd2664a9ce5de444f75dc946fae because it caused a
bug. When visiting an item type page, if the user clicked the device
breadcrumb, the url would update but the UI would not change (still look
like the item type page). Remounting on every re-render prevents this bug.

An alternative solution to the redirect bug is to not drop the item type
from the slug if it's the first render and the item type is present in
the slug but absent in the routestate. I still don't know _why_ item
type is missing from routestate.