Add condition to check if we have a situation that leads to an empty facet state.
This fucniton returns an empty list in 2 situations:
1] When the search resulted in no matching products
2] When filtering through items and there was no situation where hitsCount != item.Count

Therefore, this commit adds a check for when filteredItems is empty and adds the items that are at hitsCount === item.Count.
   - This takes care of any situation where these 2 values are the same and since we want to display grayed out facets we need
these items now.
Based on this boolean value, we will be able to toggle the `isDisabled` value for AccordianItem which does the graying out.