Disable the facets when all facets are hidden
In this commit, we check to see if for every facet, is `isHidden` true.
The issue describes a situation where when all the facets where hidden
we were left with the empty state. We now check to see if it is the case
that all facets are hidden then don't hide them but instead make `isDisable`
true which will gray out facets.