fix: Missing slash in url
The url's didn't have a trailing slash, and the path the url was
concatenated with in the Matomo script didn't have a leading slash, so
the result was a bad url.

Let's add a leading slash to the path. If for some reason the url's have
a trailing slash (or don't) in any environment, the link will still be
valid.