fix: Use plain script, put in next/head
The Script component put the script in the body, and we want it in the head.
Turns out you can't nest the Script component under the Head component, Next
throws errors. Next asks that you give each element under the Head a `key` so
that it only renders it once.