Fix the missing productlist flairs
Previously, the flairs were missing when the product was out ot stock
or when it wasn't for sale due to the `&&` consition which is
removed in this commit. Now those labels will always be visible
if the product has a lifetime warranty and/or if it is a genuine part.