origin: Use absolute & relative ifixit origin
We want to use relative origins on API requests so we don't get blocked
by CORS on previews.

We'll use absolute urls everywhere else, particularly when linking to
other pages. If we try to use a relative url with the JS URL()
constructor, it will throw an error, so we need to use absolute links in
those cases.