Fix the highlight that was getting cut off
This commit fixes it by adding some padding and then reducing
yhe space within the grid accordingly.
The grid space is altered because when the padding is added,
it moves the product model buttons down along the y axis
and they would get cut off by the show more toggle.
So by adding the padding and then modifying the spacing
we move things arounf slightly to achieve what we want.