InstantSearch: Render facets with URL from environment on the server
We were having problems where the HTML page returned from the Vercel
SSR included facet links pointing at the Vercel subdomain where it's
actually hosted, instead of at `www.ifixit.com`. This switches things
around to use the host from `NEXT_PUBLIC_IFIXIT_ORIGIN` when rendering
on the server, which means that we're in full control of where the
code thinks its origin is.

I didn't handle the case where `NEXT_PUBLIC_IFIXIT_ORIGIN` has a path
on it; it doesn't seem currently relevant, and it's kinda a change
from the previous behavior (which breaks up the `location` and uses it
to generate the new URL; the path is handled separately).