breadcrumbs: Don't use `hidden` and `display`
There is no functional change here. I just found the use of `hidden` and
`display` on the same component a little hard to read, and I thought
using a ternary in `display` instead would be a little more clear.

The intent is to hide this second breadcrumbs bar on desktop in all
cases, and only show it on mobile if we're not on /Parts or /Tools.