Fix the facets displayed on `/Tools` and children of Tools.
Previously, the children of Tools still displayed all the facets
that were hidden on `/Tools`. This commit fixes that by checking
if the productList type is AllTools or ToolsCategory and in both cases
we exclude the facets under excludedToolsFacets from being displayed.

This commit also adds Part or Kit to the exclude list.