ssr: Use device item type from slug
Modifies the `useDevicePartsItemType()` hook to use the item type from
the slug on the server.

Background:

Algolia's `useCurrentRefinements()` hook doesn't include the 'Item Type'
filter on page load, even when we visit an item type page like
/Parts/iPhone/Batteries. This is unexpected because we explicitly add
the 'Item Type' to the current UIState in the InstantSearchProvider.
This approach is a workaround for this unexpected behavior.