strapi: Upgrade 4.1.8 -> 4.1.10
We want to ultimately upgrade to 4.3.0, but Strapi has a migration guide
for this intermediate migration.
https://docs.strapi.io/developer-docs/latest/update-migration-guides/migration-guides/v4/migration-guide-4.1.8-to-4.1.10.html

I don't think we need to run the database migration script they provide
because it is only applicable if you have files in the Strapi DB whose
MIME type is null. I couldn't find any files in strapi.ifixit.com, let
alone media whose MIME type was null, so I don't think we need the
script.