routing: Tolerate case insensitive handles
This changes our product list lookup in Strapi to be case insensitive
for all pages. If we find a matching product list, then we'll use the
idiomatic `deviceTitle` from that product list throughout the rest of
our code, instead of the potentially lowercase handle in the url.

Consequently, this will add canonicals to pages with lowercase handles,
which is desirable from an SEO perspective.