MetaTags: Use strapi metaTitle if it exists
If there is a custom `metaTitle` set for this collection in Strapi, use
it in the `<title>`.

This leaves the product titles used in the rest of the app unchanged.
This can cause the title shown in the browser tab to not match the h1
title shown on the page.