logAsync: fix bug around promises
It turns out I missed something in the spec. promise.finally() will create a
*new* Promise with the same result as the underlying promise. Because
it's new and it has the same result, a new rejected promise will be
created and thus will end up being thrown because it's unhandled
(there was no .finally().catch())

However, returning promise.finally() effectively just allows the code to
run when the promise is resolved but otherwise pass the promise straight
through.