GA: send collection pageviews
On www.ifixit.com, we get pageviews in the Google Analytics dashboard. Without this line, collection pages weren't sending the pageview event. It's strange that the code on www.ifixit.com doesn't have this line in it's GA script, which is probably why it was omitted in the first place.