sentry: Report `fetch()` failures
Sentry only captures unhandled errors. `fetch()` doesn't throw for HTTP status errors, so we need to manually capture those events/log them.

In the event of a network/CORS issue, `fetch()` will throw. We probably don't want to hear about most of these, so I opted to log them but not report them to Sentry.

This also fixes an issue where we always added the request data to Sentry's global scope, whether the request succeeded or not. This only adds it to the scope when there is an HTTP error.