Fix the NewsLetter prompt movement on subscribe
When the user would subscribe, the way the different parts of that
footer would re-adjust was causing the "Stay in the loop" prompt
to move relative to the change in space and that didn't look
very good. This commit:
- Displays the text per usual until subscription status is subscribed
- Once the status is subscribed, padding to the right is added to
bring the text back to the original position after the shift.
These changes keep the position of the text consistent and does not
result in a shift after user is subscribed.