Control movement based on subscribe status
In this commit:
- The reason the text was moving on subscription was because
we were hiding the button. Here we set visibility to `hidden`
so that the button is there there but just invisible.
  - This keeps the size of that container consistent and prevents the
shifting.
- The reason the direction is being set to row or column based on
subscription status is because if `base` isn't set to column then
when we make the button invisible, the `Subscribed!` text is weirdly
on the left on tab and mobile with an invisible button next to it.
Moving the button below that text centers the text and moves the invisible
button below the text.