setSentryPageContext: Accept `context` param
Accepts the same param that is passed to `getServerSideProps` so that we can easily call it in each file in /frontend/pages.

Clears the Sentry scope before setting the tag, because this is going to be called on a JS server that handles many requests, and we don't want to accidentally include context from another request in a Sentry event.

Bumps @sentry/nextjs to 7.8.1, which is what we use in /frontend.

Uses the react-library tsconfig.json, because TS was complaining that we imported `urlFromContext()` from a .tsx file.