Fix the button position on error
Since we have a wrapper now, the error message needs takes up space
after the stack and causes a weird button position. Here I set
the error message height to 2 and then add that much to the bottom of the button
so we can avoid the weird positioning. This change is only for when we have
a subscription error. The regular behavior of everything else is thesame.