Remove onClose
In https://github.com/iFixit/react-commerce/commit/e905db9176b532d461d469711a0e3e0f64e49dbb,
we made it so that the on mobile view, when we select a facet in the
single select facet category, we close the drawer and get back to the
prodyct page. The issue with that is if the user needs to filter further,
they have to repeat certain steps.

Instead, lets keep the behavior so that when the user selects a facet
in the single select category, we bring the user back to the facet list
instead of closing the drawer.