debug: Report when device handle is `data:image`
We're seeing fetch's sent to our device wiki api that have a device of `data:image`.
https://sentry.io/share/issue/2fb1c3733997425ea19313d477f3d28e/

This is likely a bug in our frontend code, because we shouldn't be making requests like that. This will give us more "context" (pun intended) about what's causing the problem by creating a Sentry event with all the context info.