Footer: Split into Guide and Cart footers
The first step is to split our current footer into Cart and Guide
specific footers. While a part of this pull will be to make the
'Guide' footer work on both cart and guide pages, I do not want
to step on any toes yet by changing the cart footer.

By the end of this pull, we should be ready to make minimal changes
to use the 'Guide' footer everywhere whenever the Cart team is ready.