Guide Footer: add hardcoded data
I am making this new footer so that we have a footer that is not
completely reliant on cart-based props being passed in. For now,
we are going to use hardcoded data until we wire in Strapi.

Note: this data looks a bit different than the current footer. I
followed the Figma for the new footer which has a slightly
different layout.