Partners: Update styling
Now that this section is in a new place, we need to update the
styling a bit. We aren't restricting it to only ever show 3 columns,
so I added a minChildWidth to have it adjust the columns automatically,
but with a maxWidth to enforce 3 columns until we reach the large
breakpoint.