ifixit/ui/footer: split guide and cart footers
So, I goofed. This whole time I was trying to not mess with the current
footer at all, then I went and changed the styling for all of the
components. For my code reviewers, here's what this commit does (since
it probably looks like I changed everything):
   1. Moved all the components I have changed thus far into
      /guide/components
   2. Copied all of the components from main into /cart/components

Overall, the changes I made in this pull _should_ make it super simple
to merge them together, but that will be tackled later.