fix(redirect): Escape item type mappings
Fixes a bug where we forgot to escape mappings of product tags to item types. In practice, we didn't have any problems because we don't currently have any mappings that need to be escaped, but this is certainly more correct.

Also, renames `slug` -> `device`, since it's a little more descriptive and the penultimate path part is not the slug.