react-instantsearch-hooks-server: apply patch
We have an open pull to make a change in this algolia repo, but they
likely won't merge it. Since we need this diff to build on vercel, let's
just include it directly.

When we upgrade to React 18 we just need to drop the `.js` in the import
statement.

This was made using pnpm's built-in patch commands. Note: `pnpm patch`
just hung on pnpm v6, so I updated to v7, ran the commands, then
downgraded back to v6.