InstantSearchProvider: drop url hacking in createUrl()
This code is no longer needed. It *was* using the current url's ItemType
path component and tacking it on the new Url if it didn't have an
ItemType.

When looking at `Parts/Mac/Batteries` and clicking `/Parts/Macbook`,
this code was tacking on `Batteries` to the url (and not updating the
search state).

I'm not 100% sure why we needed it in the past (and only on the first
render). But the new version of instantsearch-hooks seems to not need
this. The url now is solely dependent on the state.