InstantSearchProvider: remount on url change
next.js history tracking (pushState/popState) and algolia's still don't
play nice together. In the past, we've solved this by causing the
`InstantSearch` component to remounted on each re-render (using a
renderCount state). This would *generally* only happen when the parent
component was re-rendered. That generally only happened when next.js
updated the Page in place with new props after clicking a next/link.

A more direct way of getting this done is to set the `key` prop to
something that only changes when next.js re-renders; the url.

This url prop comes from next.js and only changes when next's
knowledge of the url changes (clicking a next/link) which is exactly
when we want to remount `InstantSearch`.