Sentry: don't report 404s
If someone visits /Parts/made-up-device-name and this API 404s, let's
not bother reporting this to Sentry.

Previously we were curious if this would help us find bugs (linking to
pages that don't exist). But mostly we just get botspam that's just
munging our urls to see what happens.

I've looked at the last 20 or 30 and found many variants on the same
thing (Random edits to urls, urls for devices that don't exist, messe up
punction in the urls, ...). None of the errors made me think we were at
fault, so let's stop being hassled.