pnpm: install v7 in our GH actions
We upgraded the version of pnpm that we use for the repo, but I missed
these literal version refrences.

Let's not pin the version because I doubt vercel pins the version of
pnpm they use (they just say they will use v7 if your pnpm lock file
format indicates so).