useDevicePartsItemType: Drop SSR hack
This was added a bit ago because the server side state for algolia
wasn't matching up (or wasn't available).

See: a5481b5c597546946dd30bb5c24bf05bdfcdae9b

This concern has been addressed and this code is now *causing* a
serverside / clientside state mismatch and react is complaining that the
two states don't match. Removing this and depending on algolias
knowledge of the refinements produces relaible results.