refactor: noindex previews via middleware
Instead of noindexing dev domains in getServerSideProps, we can noindex all pages via a middleware.

We tried deploying this in #771, and it started noindexing www.ifixit.com. Turns out that the host was always react-commerce-prod.vercel.app because that's what CloudFront sets.

Instead of checking the host (which will always be `react-commerce-prod.vercel.app` on www.ifixit.com), let's check the `User-Agent`, which will be `Amazon CloudFront` for www.ifixit.com.
https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/RequestAndResponseBehaviorCustomOrigin.html