fix: noindex dev domains in code, not middleware
Middleware is run in a different location than our app. It runs on every cache hit, which is slow. We also noticed that the number of requests to our app spiked 5x and Strapi 503s increased by 20x

We're not sure why these things are happening, but for now let's do this in code instead.