fix(analytics): consolidate add to cart events
Fires one event per add to cart action, even if it was a bundle action. This prevents duplicate add to cart events in Google Analytics.

On ifixit.com, add to cart analytics events (specifically on Google Analytics) include all itemcodes in the event action. So for the cross sell section, all items that were added to the cart should show up in one event.