User API: cache response in local storage
Previously, there were some UI components that couldn't be drawn
correctly until we heard back from the user API about the pro status of
the given user. This meant either showing the wrong UI or hiding the UI
until the response came back.

No more! This will still happen on the first page load, but after that,
we store the user api data in localStorage (if available). We then use
that value as the "placeholder data" for react-query. This is
effectively used as the data *before* the user api response comes back.

99% of the time, this wil get us the right data on the first render
pass. In that 1% of the time when the data is out of date (user has
logged out or logged in just now), the cache will update as soon as the
response comes back.

We also memoize the retrieval of the value from the storage so it's only
done once per component (not once per render).

This fixes the flash of the non-pro price on first render for pro users.

Co-Authored by @sterlinghirsh