fix(cypress): drop unnecessary `wait()`'s
Cypress considers overuse of `.wait()` an antipattern.
https://docs.cypress.io/guides/references/best-practices#Unnecessary-Waiting

It makes for slower tests, and it can be flaky if the wait wasn't long enough.

The testing library we're using (the findBy.. functions) will keep retrying until the elements appear on the DOM, so let's just depend on that and drop the wait()s.
https://testing-library.com/docs/cypress-testing-library/intro

This also changes the 'cart-drawer-remove-item' click to be more idiomatic. What we had before didn't seem to click the right elements after dropping the wait().