fix(pricing): show lowest price available to user
This shows the lowest price available to a user on the product page.

If the discount is lower, that is shown and no pro badge is displayed.

If the user is a pro/employee and that price is lower, the pro price is shown. The pro badge also is shown, unless the user's discount tier is employee.