Other Stores: Add international buy box
This should function the same as it did on the previous product page.

* Makes an API request to ifixit.com from the browser for the given
  product
* That responds with "user is geolocated in country X, here's info for a
  banner for each of the variants"
  * OR it responds with null
* We use a react component to render the "Buy from our X Store" banner
  for the currently selected variant
* Shows a "or buy from the US Store" link that will dismiss the buy box
  for 7 days
   * The previous product page used a 7-day cookie, so I kept it the
     same.

Note: The UI differs from the old product page bannerl slightly and we
can definitely tweak it, but I might do that in a follow-up pull unless
there are bugs.

Also, matomo events will be in a follow up pull.