Value Proposition: hide iFixit's for drop-shipped items
We don't want to say all the "lifetime warranty" stuff for products
that are dropshipped and thus not actually guaranteed / warrantied by
us.

Note: I'd rather have this conditional on variant.fulfillmentService,
but that info is literally not availble from the graphQL api. We set
that value at the same time and from the same source as this meatfield,
so they *will* stay in sync.