IntlBuyBox: call hook once and pass down props
Derp! The Hook of course has it's own state. It loads its state from
local storage, but updates aren't broadcasted to other instances of the
hook. So, instead we call it higher up and pass the information down
where needed.

A global state store that allows multiple consumers (like redux) sounds
like a good approach, but this also works.