feat(RefinementSingleSelect): render all itemtypes
Render all item type links, and hide the ones that overflow. We want to have all the links in the DOM on page load for SEO reasons.

The refinement list `limit` must be less than the `showMoreLimit`, otherwise Algolia throws an error that says "Error: `showMoreLimit` should be greater than `limit`.". This is why I included the `- 1`.