OEM: make text work well in a sentence.
The strings look like "Geunine HTC Vive Part" so the capitalization
isn't appropriate to use in a sentence:

   This is a Genuine HTC Vive Part

These text bits come from an Akeneo SimpleSelect attribute. The values
are used elsewhere as badges so we shouldn't change them willy nilly.

Here, we lowercase the first word and Tool/Part and then add a period.

   This is a genuine HTC Vive part.