fix(algolia): don't strip query params
Fixes a bug where query params on collection pages that weren't used by Algolia were stripped from the url.

This keeps all query params in the url, even if they aren't used in Algolia search.

This means we don't have to manually forward the `_vercel_no_cache` param anymore.

I also noticed that we can pass an option to `qs` to ignore the leading '?' in the query, instead of slicing it out ourselves.
https://github.com/ljharb/qs