Checkout: Unbreak checking out again after purchase
We have code that will try to use the shopify checkoutid frmo
localstorage. This code also seems to deal with the error state of that
checkoutid having already been completed or just plain invalid.

The code is written to the spec of the API docs and appears to do that
correctly. The API however, doesn't behave as the docs say.
Specifically, when these errors are encountered, the error messages come
back out of band:

Expected response:

   {
      data: {
         checkoutLineItemsReplace: {
            checkout: { ... },
            userErrors: [{code: "ALREADY_COMPLETED"}]
      }
   }

Actual response:

   {
      data: {
         checkoutLineItemsReplace: null
      },
      errors: [
         {
            "message":"Checkout is already completed.",
            "locations":[{"line":3,"column":7}],
            "path":["checkoutLineItemsReplace"]
         }
      ]
   }

Anyway, I've checked the other endpoint here (checkoutCreate) and it has
the same failure mode (errors come out of band and look different than
the docs say). Sometimes the top level "data" property is event absent.

We may need to catch this situation deeper in the stack, but that will
come separately.