use-checkout: be more defensive
Using the locally store checkout id is a nice-to-have, but it's not
required and we certain never want to break checkout cause it fails.

We have a backup option (create a checkout) and if anything goes wrong
we should use that. So, catch all the problems and fall back.

Previously, using the locally stored checkout id sometimes threw an
error and that would break checkout. No longer!