fix(collections): skip SSR hack for breadcrumbs
We want to use the item type slug on the server for the collection page title, among other things.

However, we see a strange bug with breadcrumbs where the last breadcrumb is ellipsis on desktop, and the ellipsis are hidden on mobile, which is the opposite of what we want.

We think this is because the item type is _correct_ on the server (because of the ssr hack), and initially the item type is absent on the client. After a few renders, the item type is correct, but it puts the breadcrumbs in a weird state.

Relevant commit that addressed the same problem a while back: https://github.com/iFixit/react-commerce/commit/4ec07e2e3fdfd4c415742acfac6202d8ac94257d