fix(CORS): use relative urls in dev
When we're proxying our Next app on cominor, use relative urls for external links and api requests. This will prevent CORS errors in dev, and keep developers on the same domain when clicking on links.

The only place we are still using a hard coded origin is in frontend/pages/_document.tsx, since I don't think we have access to the request context in that file.

If the app is being visited at *.vercel.app or ifixit.com, then the IFIXIT_ORIGIN value will be used as before, so there should be no change in behavior.